+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
 
 ### 🐛 Fixes
-- [6759](https://github.com/vegaprotocol/vega/issues/6759) - Send events when liquidity provisions are undeployed
+- [6759](https://github.com/vegaprotocol/vega/issues/6759) - Send events when liquidity provisions are `undeployed`
 
 
 ## 0.62.0